Separated project into data structure, api endpoints, session handling, and a common package
