+++ src/main/java/com/thiccWallet/FCL/data/coin/Coin.java
@@ -1,4 +1,4 @@
-package com.thiccWallet.FCL.coin;
+package com.thiccWallet.FCL.data.coin;
 
 import javax.persistence.*;
