+++ src/main/java/com/thiccWallet/FCL/data/coin/CoinController.java
@@ -1,4 +1,4 @@
-package com.thiccWallet.FCL.coin;
+package com.thiccWallet.FCL.data.coin;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
