+++ src/main/java/com/thiccWallet/FCL/data/coin/CoinService.java
@@ -1,4 +1,4 @@
-package com.thiccWallet.FCL.coin;
+package com.thiccWallet.FCL.data.coin;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
